hire-api/pyproject.toml
[tool.poetry]
name = "peotry-app"
version = "0.1.0"
description = ""
authors = ["priya1008 <priyapatil20191l@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "2.0.6"
celery = "4.2.0"
coreapi = "2.3.3"
ddtrace = "0.16.0"
django-admin-rangefilter = "0.3.7"
django-extensions = "2.1.5"
django-filter = "2.0.0"
django-import-export = "1.0.1"
django-model-utils = "3.1.2"
django-reversion = "3.0.0"
djangorestframework = "3.8.2"
djangorestframework-gis = "0.13"
drf-url-filters = "0.5.1"
elasticsearch = "6.3.0"
gunicorn = "19.8.1"
marshmallow = "2.15.3"
psycopg2-binary = "2.7.4"
raven = "6.9.0"
redis = "2.10.6"
requests = "2.20.0"
splitio-client = "6.2.2"
Pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
